requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24
tqdm>=4.65

llama-cpp-python>=0.1.59

transformers>=4.35.0
accelerate>=0.22.0
safetensors>=0.3.1
torch>=2.0.0

python-dotenv>=1.0.0